Fix for https://jdic.dev.java.net/issues/show_bug.cgi?id=152
Description:
When an AWT component is added to a container, the container will call the component's addNotify(). And when the container wants to remove the component, the removeNotify() of the component will be invoked. The default implementation of java.awt.Component.removeNotify() will release its native screen resource. When the component is added to the container again or to another container, the component's addNotify() is invoked again. And addNotify() will create the native screen resource again. In our case, say WebBrowser, the removeNotify() will release all the native resource including the native browser window which is created in our native process (uta: we are loosing web page position, authentication session and etc.). So, when the addNotify() is invoked again, there is nothing, but a blank Canvas which is created by the addNotify().

Implemented (backported) solution:
WebBrowser.addNotify(), WebBrowser.removeNotify(), and WebBrowser constructors were rewritten. Old functionality stills the same. So, the old applications would have the same problem while containers change, except the restoring previous URL in browser (it solves the blank screen problem, but scroll position and authenticated session still gone). But after the fix WebBrowser component could be created with autoDispose=false parameter. In this case removeNotify() just invokes setVisible(false) and developer becomes responsible for WebBrowser.dispose() ( super.removeNotify() ) call at the finish. WebBrowser.addNotify() in this case invokes super.addNotify() only if object was not earlier initialized and calls setVisible(true) otherwise.

As mentioned, I added dispose() method to WebBrowser to release the native screen resource including the native browser window. This method should be called by user when he/she never need this instance of WebBrowser.